requirements.txt
discord.py
time
random